Created a base class for palette entry property drawers.
